+++ src/spec/vitest/QueryManagerSpec.ts
@@ -49,6 +49,39 @@ describe('A QueryManager', () => {
     expect(queryAB.getEntities(), 'No entities should match').toEqual([]);
   });
 
+  it('can create tag queries for entities', () => {
+    const world = new ex.World(null);
+    const entity1 = new ex.Entity();
+    entity1.addTag('A');
+    entity1.addTag('B');
+
+    const entity2 = new ex.Entity();
+    entity2.addTag('A');
+
+    world.entityManager.addEntity(entity1);
+    world.entityManager.addEntity(entity2);
+
+    // Query for all entities that have type A components
+    const queryA = world.queryTags(['A']);
+    // Query for all entities that have type A & B components
+    const queryAB = world.queryTags(['A', 'B']);
+
+    expect(queryA.getEntities(), 'Both entities have component A').toEqual([entity1, entity2]);
+    expect(queryAB.getEntities(), 'Only entity1 has both A+B').toEqual([entity1]);
+
+    // Queries update if component change
+    entity2.addTag('B');
+    expect(queryAB.getEntities(), 'Now both entities have A+B').toEqual([entity1, entity2]);
+
+    // Queries update if components change
+    entity2.removeTag('B');
+    expect(queryAB.getEntities(), 'Component force removed from entity, only entity1 A+B').toEqual([entity1]);
+
+    // Queries are deferred by default, so queries will update after removals
+    entity1.removeTag('B');
+    expect(queryAB.getEntities(), 'No entities should match').toEqual([]);
+  });
+
   it('can add entities to queries', () => {
     const world = new ex.World(null);
     const entity1 = new ex.Entity();
@@ -111,6 +144,97 @@ describe('A QueryManager', () => {
     expect(queryAB.getEntities()).toEqual([]);
   });
 
+  it('can remove entities from queries that have components and not(tags)', () => {
+    const world = new ex.World(null);
+    const entity1 = new ex.Entity();
+    entity1.addComponent(new FakeComponentA());
+    entity1.addComponent(new FakeComponentB());
+    entity1.addTag('ex.offscreen');
+
+    const entity2 = new ex.Entity();
+    entity2.addComponent(new FakeComponentA());
+    entity2.addComponent(new FakeComponentB());
+
+    const queryAB = world.query({
+      components: {
+        all: [FakeComponentA, FakeComponentB]
+      },
+      tags: {
+        not: ['ex.offscreen']
+      }
+    });
+    world.queryManager.addEntity(entity1);
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([entity2]);
+
+    entity2.addTag('ex.offscreen');
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([]);
+
+    entity1.removeTag('ex.offscreen');
+    world.queryManager.addEntity(entity1);
+    expect(queryAB.getEntities()).toEqual([entity1]);
+  });
+
+  it('can remove entities from queries that have components and not(components)', () => {
+    const world = new ex.World(null);
+    const entity1 = new ex.Entity();
+    entity1.addComponent(new FakeComponentA());
+    entity1.addComponent(new FakeComponentB());
+    entity1.addComponent(new FakeComponentC());
+
+    const entity2 = new ex.Entity();
+    entity2.addComponent(new FakeComponentA());
+    entity2.addComponent(new FakeComponentB());
+
+    const queryAB = world.query({
+      components: {
+        all: [FakeComponentA, FakeComponentB],
+        not: [FakeComponentC]
+      }
+    });
+    world.queryManager.addEntity(entity1);
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([entity2]);
+
+    entity2.addComponent(new FakeComponentC());
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([]);
+
+    entity1.removeComponent(FakeComponentC, true);
+    world.queryManager.addEntity(entity1);
+    expect(queryAB.getEntities()).toEqual([entity1]);
+  });
+
+  it('can add entities queries that have 1 and(component) + or(components)', () => {
+    const world = new ex.World(null);
+    const entity1 = new ex.Entity();
+    entity1.addComponent(new FakeComponentA());
+    entity1.addComponent(new FakeComponentC());
+
+    const entity2 = new ex.Entity();
+    entity2.addComponent(new FakeComponentA());
+    entity2.addComponent(new FakeComponentB());
+
+    const queryAB = world.query({
+      components: {
+        all: [FakeComponentA],
+        any: [FakeComponentB, FakeComponentC]
+      }
+    });
+    world.queryManager.addEntity(entity1);
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([entity1, entity2]);
+
+    entity1.removeComponent(FakeComponentC, true);
+    world.queryManager.addEntity(entity1);
+    expect(queryAB.getEntities()).toEqual([entity2]);
+
+    entity2.removeComponent(FakeComponentB, true);
+    world.queryManager.addEntity(entity2);
+    expect(queryAB.getEntities()).toEqual([]);
+  });
+
   it('can remove entities from tag queries', () => {
     const world = new ex.World(null);
     const entity1 = new ex.Entity();
@@ -150,7 +274,7 @@ describe('A QueryManager', () => {
     expect(queryAB.getEntities()).toEqual([entity1, entity2]);
 
     const removed = entity1.get(FakeComponentA);
-    entity1.removeComponent(FakeComponentA);
+    entity1.removeComponent(FakeComponentA, true);
     world.queryManager.removeComponent(entity1, removed);
 
     expect(queryAB.getEntities()).toEqual([entity2]);
@@ -170,12 +294,16 @@ describe('A QueryManager', () => {
     world.queryManager.addEntity(entity1);
     world.queryManager.addEntity(entity2);
 
-    expect(queryAB.getEntities()).toEqual([entity1, entity2]);
+    expect(queryAB.entities).toEqual([entity1, entity2]);
 
     entity1.removeTag('A');
     world.queryManager.removeTag(entity1, 'A');
 
-    expect(queryAB.getEntities()).toEqual([entity2]);
+    expect(queryAB.entities).toEqual([entity2]);
+
+    entity2.removeTag('B');
+    world.queryManager.removeTag(entity2, 'B');
+    expect(queryAB.entities).toEqual([]);
   });
 
   it("removing components unrelated to the query doesn't remove the entity", () => {